+++ src/plugins/pcr/__init__.py
@@ -4,20 +4,19 @@
 
 参考 https://github.com/pcrbot/yobot
 """
-from nonebot import CommandGroup
+from nonebot import on_command
+from nonebot.plugin import PluginMetadata
 
 from .data import calendar
 
-pcr = CommandGroup("pcr")
+__plugin_meta__ = PluginMetadata(
+    name="公主连结Re:Dive",
+    description="日程表查询",
+    usage="获取一周日程表\n/公主连结日程表",
+)
 
 # region 日程表
-calendar_cmd = pcr.command("calendar", aliases={("pcr", "日程表"), ("pcr", "日程")})
-calendar_cmd.__doc__ = """
-公主连结Re:Dive 日程表
-
-获取一周日程表
-/pcr.calendar
-"""
+calendar_cmd = on_command("公主连结日程表")
 
 
 @calendar_cmd.handle()