+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/lang/zh-CN/
 - 添加数据库插件
 - 添加词云插件
 - FFLogs 添加设置缓存的命令
+- 适配插件元信息
+- 使用 nonebot-plugin-treehelp 生成插件帮助
 
 ### Changed